+++ CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+- BehaviourStates property and Get functions into AnimatorStateMachine component.
+
 ## [1.1.0] - 2022-12-11
 ### Added
 - GenericMonoBehaviourState